Split tool wrappers into individual modules
